Packet 12: a handle names one token, and the resolve path stops saturating
Handle issuance refuses instead of saturating. The text reader's `issue` took
the table's length, converted it with a saturating fallback, and handed back
`SpanHandle::at(index)`. Past the width a handle is carried in, every issue
answered with the same index: two tokens holding one handle, and a diagnostic
resolving to whichever the table reached first. That is a wrong answer wearing
the shape of a right one, which is the class the plane refuses by name. `issue`
now returns a refusal at the byte the token sits at, and its five call sites
carry it.

The resolve path composes its refusal instead of saturating twice. Looking up a
handle narrowed the handle's index with one saturating conversion and reported
how far the table reaches with a second. Neither is a display: an index that
saturated indexes the wrong row, and a reach that saturated states a width the
table does not have.

`SpanResolutionRefusal::reaches` is a length, so it is a `usize`. It was a `u32`
reached through a conversion that could only be honest by luck. Nothing derives
an identity from it and no encoder writes it - its one reader is the sentence
`described()` renders - so the width was carrying a conversion rather than a
meaning.

Four saturating conversions removed from the token home. The four that remain
there are `usize` to `u64` widenings, which no supported target can fail, and
the issued-count scalar that Packet 12 settles separately by deriving the count
from the table rather than accepting it beside one.

+++ macros/macroc/src/token/types.rs
@@ -222,7 +222,7 @@ pub struct SpanResolutionRefusal {
     pub handle: SpanHandle,
     /// How many positions the table carries.
     /// A handle at or past this index names no position in it.
-    pub reaches: u32,
+    pub reaches: usize,
 }
 
 /// How a producer answers "where is the token this handle names?".
